requirements.txt
fastapi
uvicorn
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pydantic
python-dotenv
